requirements.txt
django==4.0
djangorestframework==3.13.0
requests
django-cors-headers
psycopg2-binary
python-dotenv
gunicorn
whitenoise
dj-database-url
django-cors-headers
drf_yasg